Revert "Demonstration of how to Backpack'ify this package."
This reverts commit de314fd0fc32110c0ca2a64bfeb2a5def5806c5d.
