--- types/Parser/Text.hs
@@ -1,7 +1,0 @@
-module Parser.Text (module Parser.Text, module Data.Attoparsec.Text) where
-
-import Data.Text
-import Data.Attoparsec.Text
-
-takeRest :: Parser Text
-takeRest = takeText
